Thanks to @juanfrilla for https://github.com/FlareSolverr/FlareSolverr/issues/1036#issuecomment-2066119004.

Unfortunately, currently this only works on Windows, and the looping challenges return if using proxies or VPNs.