@juanfrilla 2024-04-21
Another thing that I've noticed is that in the user-agent headless replacement:
```python
                self.execute_cdp_cmd(
                    "Network.setUserAgentOverride",
                    {
                        "userAgent": self.execute_script(
                            "return navigator.userAgent"
                        ).replace("Headless", "")
                    },
                )
```
I don't know why but If I hardcode the user-agent using the exact that my computer has like this:
```python
user_agent = "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/123.0.0.0 Safari/537.36"
options.add_argument(f"--user-agent={user_agent}")
```

 it works, but if i put this to make it automatically like the first example it does not work.
 
 
 PD: I only can guide through the solution cuz I'm trying to get the project installed/set up 😅