@ilike2burnthing 2024-04-21
@garfield69 yea this seems to be an issue with Chrome v124. You can revert to v123 in the mean time if it's easier - https://github.com/FlareSolverr/FlareSolverr/issues/1161

Alternatively build your own binaries:

> python build_package.py