@ilike2burnthing 2024-04-21
@m33ts4k0z were you doing this on Windows?