@garfield69 2024-04-21
>Alternatively, build your own binaries, which will use Chromium v123:

>>python src/build_package.py

Oh cool, did not know I could build on windows.
Built successfully and tested. Much better, no left over chrome tasks chewing CPU anymore :-)