@ilike2burnthing 2024-04-22
@juanfrilla sorry for the delay in replying, been busy an only got to few quick ones on my phone.

I'll have a look at the UA idea when I next get a chance, thanks.

Assuming you're following the run from source instructions, what issue are you having? https://github.com/FlareSolverr/FlareSolverr#from-source-code