@juanfrilla 2024-04-22
@ilike2burnthing my main problem is that i cannot install Xvfb on MacOS