@ilike2burnthing 2024-04-22
Tried XQuartz?