@juanfrilla 2024-04-22
> Tried XQuartz?

yessir now the project is set up, let's try what I can fix