@juanfrilla 2024-04-22
Locally it works for my website:
```python
import requests
import json

url = "http://localhost:8191/v1"
headers = {"Content-Type": "application/json"}
url = "https://www.icj-cij.org/sites/default/files/case-related/187/187-20231215-ord-01-00-en.pdf"
data = {"cmd": "request.get", "url": url, "maxTimeout": 60000}
response = requests.post(url, headers=headers, json=data)

with open("output.pdf", "wb") as file:
    file.write(response.content)
```

But i want the cookies here and I cannot access to `response.json()`