@ilike2burnthing 2024-05-01
> Unfortunately, currently this only works on Windows, and the looping challenges return if using proxies or VPNs.

> I'll have a look at the UA idea when I next get a chance, thanks.