@21hsmw 2024-05-01
> > Unfortunately, currently this only works on Windows, and the looping challenges return if using proxies or VPNs.
> 
> > I'll have a look at the UA idea when I next get a chance, thanks.

Well, I made my own implementation of this "new tab" idea and I was able to make it work with every website I could (ext.to, www3.yggtorrent.cool, dodi-repacks.site, hd-torrents.me/login.php, nhentai.net) on my Linux system using a VPN / socks5 proxy and also with my container image on my own remote Linux server, which was blocked by cloudflare too.
Unfortunately I can't test on Windows, so if someone can test that and report back please do.

Public image with my edits: [21hsmw/flaresolverr:fixlooping](https://hub.docker.com/layers/21hsmw/flaresolverr/fixlooping/images/sha256-294cd52a0fba6b1079a3afe24305c36447ccd00ad440b504ffd14b2cdfe880f1?context=explore)
Code here: https://github.com/21hsmw/FlareSolverr/commit/da6cc9d1b3b6879a75da386e0da79ec491bce241