@ilike2burnthing 2024-05-02
That's working 95% of the time on Windows for me, even with a proxy, but failing 95% of the time on Docker. Usual error:

```text
Error: Error solving the challenge. 'NoneType' object has no attribute 'startswith'
```

Seems it's related to `get_correct_window` and trying to get `driver.current_url`. Adding some extra logging shows that the URL is returned as `None`. Adding some additional sleeps then shows the correct URL, but I'm still getting challenge loops or crashed.