@aevrard 2024-05-03
For yggtorrent, i resolve my issue by configure : 

    dns:
      - 8.8.8.8
      - 8.8.4.4
    environment:
      - CAPTCHA_SOLVER=hcaptcha-solver
      - LANG=en_US