@21hsmw 2024-10-31
> > If you are referring to the delay between page requests, I guess it's on fanficfare side. It's probably doing a few things after getting the page content before asking for a new one.
> 
> Im actually referring to this time ![image](https://private-user-images.githubusercontent.com/129144583/382139575-fc392ec0-93c1-406f-908e-7e1269799629.png?jwt=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJnaXRodWIuY29tIiwiYXVkIjoicmF3LmdpdGh1YnVzZXJjb250ZW50LmNvbSIsImtleSI6ImtleTUiLCJleHAiOjE3MzA0MTg0ODEsIm5iZiI6MTczMDQxODE4MSwicGF0aCI6Ii8xMjkxNDQ1ODMvMzgyMTM5NTc1LWZjMzkyZWMwLTkzYzEtNDA2Zi05MDhlLTdlMTI2OTc5OTYyOS5wbmc_WC1BbXotQWxnb3JpdGhtPUFXUzQtSE1BQy1TSEEyNTYmWC1BbXotQ3JlZGVudGlhbD1BS0lBVkNPRFlMU0E1M1BRSzRaQSUyRjIwMjQxMDMxJTJGdXMtZWFzdC0xJTJGczMlMkZhd3M0X3JlcXVlc3QmWC1BbXotRGF0ZT0yMDI0MTAzMVQyMzQzMDFaJlgtQW16LUV4cGlyZXM9MzAwJlgtQW16LVNpZ25hdHVyZT1mNDM3ZWVkMmU4MzNjNjE5M2Y1ZDZhNmU3NGRhYjBmZWYwODI5ZDllMmM1OTI4OGQ5NTFmYTZjMzBjNmNmMDFiJlgtQW16LVNpZ25lZEhlYWRlcnM9aG9zdCJ9.uZMHRfH1Cs6Rl00_h0DlJ3QaDMlHBhs02QXCGs9O_wg)
> 
> IIRC historically it was like ~1-1.5 seconds, but like... its so minor, im just glad this is working again haha

I use some hardcoded wait times between 1 and 2 seconds to allow slow systems to load pages properly before continuing. 
By removing these, I'm able to go from ~12s to ~6s resolution time and from ~3s to ~1.5s with cookies.
I already had the idea to detect the system capacity and adjust the speed accordingly, so it should be faster later on.